aoj/ALDS1/ALDS1_10_B_test.py
from ALDS1_10_B import resolve
import sys
from io import StringIO
import unittest

class TestClass(unittest.TestCase):
    maxDiff = None
    def assertIO(self, input, output):
        stdout, stdin = sys.stdout, sys.stdin
        sys.stdout, sys.stdin = StringIO(), StringIO(input)
        resolve()
        sys.stdout.seek(0)
        out = sys.stdout.read()[:-1]
        sys.stdout, sys.stdin = stdout, stdin
        self.assertEqual(out, output)
    def test_入力例_1(self):
        input = """6
30 35
35 15
15 5
5 10
10 20
20 25"""
        output = """15125"""
        self.assertIO(input, output)

if __name__ == "__main__":
    unittest.main()